refactor: update imports for ArrowUp icon component
