+++ src/presentation/globals/components/icons/Icons.tsx
@@ -1,14 +1,6 @@
 import type { IconProps } from '@/presentation/globals/presentation.types';
 import { IconWrapper } from '@/presentation/globals/components/icons/IconWrapper';
 
-export function IconArrowUp(props: IconProps) {
-  return (
-    <IconWrapper {...props}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="m4.5 15.75 7.5-7.5 7.5 7.5" />
-    </IconWrapper>
-  );
-}
-
 export function IconArrowsLeftRight(props: IconProps) {
   return (
     <IconWrapper {...props}>
